+++ packages/tailwind-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baneung-pack/tailwind-config
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @baneung-pack/tokens@1.0.3
+
 ## 0.0.3
 
 ### Patch Changes